+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 8.2.1
 ### Features
-* **validators**: new pattern validator([7eded1c](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/7eded1c))
+* **validators**: new pattern validator ([7eded1c](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/7eded1c))
 * **o-dialog**: Added classes in the buttons in dialog ([b107153](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/b107153))
+* **OFormServiceComponent** (**o-combo**, **o-list-picker**, **o-radio**): new attribute `sort`, include `translate` attribute in `o-list-picker` ([7a7032bb](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/7a7032bb)) Closes ([#322](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/322))
 * **o-table-row-expandable**: new `multiple` input ([fb4fb10](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/fb4fb10)) Closes ([#545](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/545))
 * **o-combo**: new types of render (currency,integer,boolean,percentage,date,custom render) ([eabdd2d](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/eabdd2d)) Closes ([#529](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/529))
 * **o-list-picker**: new types of render (currency,integer,percentage,date,custom render) ([eabdd2d](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/eabdd2d)) Closes ([#529](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/529))